Moved generate_token to conftest
